+++ superwall/src/main/java/com/superwall/sdk/store/abstractions/transactions/StoreTransactionType.kt
@@ -25,7 +25,7 @@ interface StoreTransactionType {
     val expirationDate: Date?
     val offerId: String?
     val revocationDate: Date?
-    val appAccountToken: UUID?
+    val appAccountToken: String?
     val purchaseToken: String
 }